extract unixfs filestore into lib
