+++ lib/unixfs/filestore.go
@@ -1,4 +1,4 @@
-package client
+package unixfs
 
 import (
 	"context"
@@ -19,12 +19,15 @@ import (
 	"github.com/ipfs/go-merkledag"
 	"github.com/ipfs/go-unixfs/importer/balanced"
 	ihelper "github.com/ipfs/go-unixfs/importer/helpers"
+	mh "github.com/multiformats/go-multihash"
 	"golang.org/x/xerrors"
 
 	"github.com/filecoin-project/lotus/build"
 )
 
-func unixFSCidBuilder() (cid.Builder, error) {
+var DefaultHashFunction = uint64(mh.BLAKE2B_MIN + 31)
+
+func CidBuilder() (cid.Builder, error) {
 	prefix, err := merkledag.PrefixForCidVersion(1)
 	if err != nil {
 		return nil, fmt.Errorf("failed to initialize UnixFS CID Builder: %w", err)
@@ -37,9 +40,9 @@ func unixFSCidBuilder() (cid.Builder, error) {
 	return b, nil
 }
 
-// createUnixFSFilestore takes a standard file whose path is src, forms a UnixFS DAG, and
+// CreateFilestore takes a standard file whose path is src, forms a UnixFS DAG, and
 // writes a CARv2 file with positional mapping (backed by the go-filestore library).
-func (a *API) createUnixFSFilestore(ctx context.Context, srcPath string, dstPath string) (cid.Cid, error) {
+func CreateFilestore(ctx context.Context, srcPath string, dstPath string) (cid.Cid, error) {
 	// This method uses a two-phase approach with a staging CAR blockstore and
 	// a final CAR blockstore.
 	//
@@ -80,7 +83,7 @@ func (a *API) createUnixFSFilestore(ctx context.Context, srcPath string, dstPath
 		return cid.Undef, xerrors.Errorf("failed to create temporary filestore: %w", err)
 	}
 
-	finalRoot1, err := buildUnixFS(ctx, file, fstore, true)
+	finalRoot1, err := Build(ctx, file, fstore, true)
 	if err != nil {
 		_ = fstore.Close()
 		return cid.Undef, xerrors.Errorf("failed to import file to store to compute root: %w", err)
@@ -102,7 +105,7 @@ func (a *API) createUnixFSFilestore(ctx context.Context, srcPath string, dstPath
 		return cid.Undef, xerrors.Errorf("failed to rewind file: %w", err)
 	}
 
-	finalRoot2, err := buildUnixFS(ctx, file, bs, true)
+	finalRoot2, err := Build(ctx, file, bs, true)
 	if err != nil {
 		_ = bs.Close()
 		return cid.Undef, xerrors.Errorf("failed to create UnixFS DAG with carv2 blockstore: %w", err)
@@ -119,10 +122,10 @@ func (a *API) createUnixFSFilestore(ctx context.Context, srcPath string, dstPath
 	return finalRoot1, nil
 }
 
-// buildUnixFS builds a UnixFS DAG out of the supplied reader,
+// Build builds a UnixFS DAG out of the supplied reader,
 // and imports the DAG into the supplied service.
-func buildUnixFS(ctx context.Context, reader io.Reader, into bstore.Blockstore, filestore bool) (cid.Cid, error) {
-	b, err := unixFSCidBuilder()
+func Build(ctx context.Context, reader io.Reader, into bstore.Blockstore, filestore bool) (cid.Cid, error) {
+	b, err := CidBuilder()
 	if err != nil {
 		return cid.Undef, err
 	}
